+++ university_scheduler/viewCoursesForm.cs
@@ -98,20 +98,31 @@ namespace university_scheduler
         private void deleteCourseBTN_Click(object sender, EventArgs e)
         {
             int selected_id = (int)courseData.SelectedRows[0].Cells[0].Value;
-            SqlConnection cn = new SqlConnection(conString);
-            cn.Open();
-            if (cn.State == System.Data.ConnectionState.Open)
+            string message = "By clicking OK, this course will be permenantly DELETED\nmake sure from your choice";
+            string title = "course will be deleted";
+            MessageBoxButtons buttons = MessageBoxButtons.OKCancel;
+            DialogResult result = MessageBox.Show(message, title, buttons, MessageBoxIcon.Warning, MessageBoxDefaultButton.Button2);
+            if (result != DialogResult.OK)
             {
-                string query = "delete from course where id = " + selected_id;
-                SqlCommand cmd = new SqlCommand(query, cn);
-                cmd.ExecuteNonQuery();
-                //--the following three lines is used to update the dataGridView and refresh it --//
-                this.loaddata();
-                this.courseData.Update();
-                this.courseData.Refresh();
-                //---//
-                MessageBox.Show("Course is deleted successfully..!");
-                cn.Close();
+                return;
+            }
+            else
+            {
+                SqlConnection cn = new SqlConnection(conString);
+                cn.Open();
+                if (cn.State == System.Data.ConnectionState.Open)
+                {
+                    string query = "delete from course where id = " + selected_id;
+                    SqlCommand cmd = new SqlCommand(query, cn);
+                    cmd.ExecuteNonQuery();
+                    //--the following three lines is used to update the dataGridView and refresh it --//
+                    this.loaddata();
+                    this.courseData.Update();
+                    this.courseData.Refresh();
+                    //---//
+                    MessageBox.Show("Course is deleted successfully..!");
+                    cn.Close();
+                }
             }
         }
